fix: フックの構成やコンポーネントの構成を変えて、あとの更新処理を書きやすくした #134
